capture, test: Move stubs to stubs_test.go
